Rename types related to L1BatchExecutorBuilder
