+++ core/lib/zksync_core/src/state_keeper/tests/mod.rs
@@ -33,7 +33,7 @@ use self::tester::{
     bootloader_tip_out_of_gas, pending_batch_data, random_tx, rejected_exec, successful_exec,
     successful_exec_with_metrics, TestScenario,
 };
-pub(crate) use self::tester::{MockBatchExecutorBuilder, TestBatchExecutorBuilder};
+pub(crate) use self::tester::{MockBatchExecutor, TestBatchExecutorBuilder};
 use crate::{
     gas_tracker::l1_batch_base_cost,
     state_keeper::{
